6.리팩토링
- 컴포넌트 js파일을 UI 계층에 따라 폴더로 분리
- 할일 상태에 따른 각각의 state list를 todoList 하나로 통합
- 불필요한 인스턴스 변수 삭제
