+++ src/App_TodoMain/TodoListView/RemoveTodoButton.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { styles } from './styles'
+import { styles } from '../styles/styles'
 
 
 // 삭제 버튼 컴포넌트
